requirements.txt
fastapi==0.111.0
uvicorn==0.30.0
sqlalchemy==2.0.30
python-dotenv==1.0.1